PMM-7574 move shared components to Elements
