+++ public/app/percona/backup/components/StorageLocations/AddStorageLocationModal/S3Fields/S3Fields.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import { TextInputField } from '@percona/platform-core';
 import { S3Fields } from './S3Fields';
-import { SecretToggler } from 'app/percona/shared/SecretToggler';
+import { SecretToggler } from 'app/percona/shared/components/Elements/SecretToggler';
 
 describe('S3Fields', () => {
   it('should pass initial values', () => {
