+++ public/app/percona/backup/components/StorageLocations/AddStorageLocationModal/S3Fields/S3Fields.tsx
@@ -3,7 +3,7 @@ import { TextInputField, validators } from '@percona/platform-core';
 import { S3FieldsProps } from './S3Fields.types';
 import { Messages } from './S3Fields.Messages';
 import { MAX_LENGTH } from './S3Fields.constants';
-import { SecretToggler } from 'app/percona/shared/SecretToggler';
+import { SecretToggler } from 'app/percona/shared/components/Elements/SecretToggler';
 
 const required = [validators.required];
