+++ public/app/percona/integrated-alerting/components/AlertRules/AlertRulesActions/AlertRulesActions.tsx
@@ -8,7 +8,7 @@ import { AlertRulesActionsProps } from './AlertRulesActions.types';
 import { AlertRulesProvider } from '../AlertRules.provider';
 import { AlertRulesService } from '../AlertRules.service';
 import { Messages } from './AlertRulesActions.messages';
-import { DeleteModal } from 'app/percona/shared/DeleteModal';
+import { DeleteModal } from 'app/percona/shared/components/Elements/DeleteModal';
 import { AlertRuleCreatePayload } from '../AlertRules.types';
 
 export const AlertRulesActions: FC<AlertRulesActionsProps> = ({ alertRule }) => {
