+++ public/app/percona/integrated-alerting/components/NotificationChannel/DeleteNotificationChannelModal/DeleteNotificationChannelModal.tsx
@@ -2,7 +2,7 @@ import React, { FC, useContext, useState } from 'react';
 import { logger } from '@percona/platform-core';
 import { AppEvents } from '@grafana/data';
 import { appEvents } from 'app/core/core';
-import { DeleteModal } from 'app/percona/shared/DeleteModal';
+import { DeleteModal } from 'app/percona/shared/components/Elements/DeleteModal';
 import { NotificationChannelService } from '../NotificationChannel.service';
 import { NotificationChannelProvider } from '../NotificationChannel.provider';
 import { DeleteNotificationChannelModalProps } from './DeleteNotificationChannelModal.types';
