+++ public/app/percona/integrated-alerting/components/Table/TableContent/TableContent.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from 'react';
-import { EmptyBlock } from 'app/percona/shared/EmptyBlock';
+import { EmptyBlock } from 'app/percona/shared/components/Elements/EmptyBlock';
 import { TableContentProps } from './TableContent.types';
 
 export const TableContent: FC<TableContentProps> = ({ hasData, emptyMessage, loading, children }) =>
