+++ public/app/percona/shared/components/Elements/DeleteModal/DeleteModal.styles.ts
