+++ public/app/percona/shared/components/Elements/DeleteModal/DeleteModal.tsx
