+++ public/app/percona/shared/components/Elements/DeleteModal/DeleteModal.types.ts
