+++ public/app/percona/shared/components/Elements/DeleteModal/index.ts
