+++ public/app/percona/shared/components/Elements/EmptyBlock/EmptyBlock.styles.ts
