+++ public/app/percona/shared/components/Elements/SecretToggler/SecretToggler.styles.ts
