+++ public/app/percona/shared/components/Elements/SecretToggler/index.ts
